Move constructor to its own file
